refactored code added proper styled components
